+++ pyaud/__init__.py
@@ -2,11 +2,10 @@
 from . import exceptions, messages, plugins
 from ._core import pyaud
 from ._main import main
-from ._objects import BasePlugin, files
+from ._objects import files
 from ._version import __version__
 
 __all__ = [
-    "BasePlugin",
     "__version__",
     "exceptions",
     "files",